requirements-app.txt
# requirements-app.txt
# Application-specific dependencies
# Versions from version.txt
# Python 3.12.x compatible

# ============================================
# WEB FRAMEWORK
# ============================================
fastapi==0.108.0
uvicorn[standard]==0.25.0
websockets==12.0
python-socketio==5.11.1
aiohttp==3.9.0

# ============================================
# DATABASE (PostgreSQL)
# ============================================
sqlalchemy==2.0.0
asyncpg==0.29.0
alembic==1.12.1
psycopg2-binary==2.9.9

# ============================================
# VECTOR DATABASE (Milvus)
# ============================================
# pymilvus==2.4.0 (already in base)

# ============================================
# OBJECT STORAGE (MinIO)
# ============================================
minio==7.2.0

# ============================================
# CACHE (Redis)
# ============================================
redis==5.0.1
hiredis==2.3.2

# ============================================
# TASK QUEUE (Celery)
# ============================================
celery[redis]==5.3.6
flower==2.0.1
kombu==5.3.4

# ============================================
# LLM CLIENT
# ============================================
ollama==0.1.6

# ============================================
# SECURITY AND AUTHENTICATION
# ============================================
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.1
email-validator==2.1.0

# ============================================
# CONFIGURATION
# ============================================
pydantic-settings==2.1.0
python-dateutil==2.8.2
toml==0.10.2

# ============================================
# STREAMLIT (Frontend)
# ============================================
streamlit==1.30.0
streamlit-authenticator==0.2.3
plotly==5.18.0
altair==5.0.1

# ============================================
# MONITORING
# ============================================
opentelemetry-instrumentation-fastapi==0.43b0
opentelemetry-instrumentation-sqlalchemy==0.43b0

# ============================================
# UTILITIES
# ============================================
click==8.1.7
rich==13.7.0
typer==0.9.0

# ============================================
# OPTIONAL: DEVELOPMENT TOOLS
# ============================================
# Uncomment these for development (from version.txt)
# pytest==7.4.0
# pytest-cov==4.1.0
# black==23.9.0
# ruff==0.1.0
# ipython==8.15.0
# jupyter==1.0.0
